% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_network}
\alias{extract_network}
\title{Extract networks from a data.frame}
\usage{
extract_network(
  df,
  g = NULL,
  vars,
  ids,
  names = get_names(id),
  net_cov,
  node_cov,
  outcome = c("network", "generic", "igraph")
)
}
\arguments{
\item{g}{A vector or column name that identifies groups in the dataset. This parameter is only needed when operating with multilevel network data.}

\item{vars}{A vector that indicates all variable that potentially contain network data. This
can either be a vector of numeric or logic indices or one of the indicator functions
\code{ends_with()}, \code{starts_with()} and \code{contains()}.}

\item{ids}{A vector or variable name that uniquely identifes each row aka participant in the dataset.}

\item{names}{An optional vector or column name that can later be used to label/identify
nodes in the extracted networks.}

\item{node_cov}{A vector or column name that}
}
\value{

}
\description{
Extract networks from a data.frame
}
